% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxLine.R
\name{boxLine}
\alias{boxLine}
\title{boxLine: Produces  two graphs - boxplots of disaggregated population counts across groups and a line plot showing the mean distribution of the
disaggregated counts.}
\usage{
boxLine(dmat, xlab, ylab)
}
\arguments{
\item{dmat}{A data frame containing the group-structured disaggregated population estimates which could be an observed data or model estimates from
'cheesecake', 'cheesepop', 'slices','spices', 'spray' , 'sprinkle', 'splash', 'spray', 'sprinkle1', 'splash1', or 'spray1'.
considered.}

\item{xlab}{A user-defined label for the x-axis (e.g., 'Frequency').}

\item{ylab}{A user-defined label for the y-axis (e.g., 'Population count').}
}
\value{
A graphic image of combined boxplot and line plots of population counts across groups.
}
\description{
This function automarically generate two graphs that are combined together - a boxplot graph (a) of the distribution of of the various group's
disaggregated population count, and (b) a line graph of the aggregated counts across all groups (e.g., total number of individuals for each group).
Its input data could come from any of the disaggregation functions within the 'jollofR' package such as 'cheesecake', 'cheesepop', 'slices', etc.
}
\examples{
data(toydata)
result <- cheesecake(df = toydata$admin, output_dir = tempdir())
boxLine(dmat=result$male_age_pop,
xlab="Age group (years)",
ylab = "Population Count")
}
