% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jollofR-package.R
\docType{package}
\name{jollofR-package}
\alias{jollofR}
\alias{jollofR-package}
\title{jollofR: A Bayesian statistical model-based approach for disaggregating small area population estimates by demographic characteristics}
\description{
Automatic dis-aggregation of small area population estimates by demographic groups (e.g.,age, sex, race, marital status, educational level, etc) along with the estimates of uncertainty, using advanced Bayesian statistical modelling approaches.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wpgp/jollof}
  \item \url{https://wpgp.github.io/jollof/}
  \item \url{https://wpgp.github.io/jollofR/}
  \item Report bugs at \url{https://github.com/wpgp/jollof/issues}
}

}
\author{
\strong{Maintainer}: Chibuzor Christopher Nnanatu \email{cc.nnanatu@soton.ac.uk} (\href{https://orcid.org/0000-0002-5841-3700}{ORCID})

Authors:
\itemize{
  \item Somnath Chaudhuri (\href{https://orcid.org/0000-0003-4899-1870}{ORCID})
  \item Attila N Lazar (\href{https://orcid.org/0000-0003-2033-2013}{ORCID})
  \item Andrew J Tatem (\href{https://orcid.org/0000-0002-7270-941X}{ORCID})
}

}
\keyword{internal}
