% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHist.R
\name{plotHist}
\alias{plotHist}
\title{plotHist: Produces histogram of the disaggregated population counts for each group}
\usage{
plotHist(dmat, xlab, ylab)
}
\arguments{
\item{dmat}{A data frame containing the group-structured disaggregated population estimates which could either be observed or predicted from
'cheesecake', 'cheesepop', 'slices','spices', 'spray' , 'sprinkle', 'splash', 'spray', 'sprinkle1', 'splash1', and 'spray1'.}

\item{xlab}{A user-defined label for the x-axis (e.g., 'Frequency')
considered.}

\item{ylab}{A user-defined label for the y-axis (e.g., 'Population count')
considered.}
}
\value{
A graphic image of histogram of the disaggregated population count
}
\description{
This function produces multi-panel histogram plots of the disaggregated population counts for each group.
Its input data can come from any of the disaggregation functions within the 'jollofR' package such as 'cheesecake', 'cheesepop', 'slices', etc.
}
\examples{
data(toydata)
result <- cheesecake(df = toydata$admin, output_dir = tempdir())
plotHist(dmat=result$age_pop, xlab="Population Count", ylab = "Frequency")
}
