% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRast.R
\name{plotRast}
\alias{plotRast}
\title{plotRast: Produces multi-panel maps of the raster files of the grid-cell disaggregated structured population counts}
\usage{
plotRast(output_dir, names, nrow, ncol)
}
\arguments{
\item{output_dir}{The raster files folder directory for the disaggregated population estimates}

\item{names}{A user-defined names for the x-axis ticks labels.}

\item{nrow}{Number of rows of the multi-panel maps. The value depends on the number of groups being displayed.}

\item{ncol}{Number of columns of the multi-panel maps. The value depends on the number of groups being displayed.}
}
\value{
A graphic image of multi-panel maps of population disaggregated raster files
}
\description{
This function produces multi-frame maps of the raster files across the various demographic groups of interest.
The input data could come from any of the jolloR disaggregation functions at grid cell levels, e.g.,'sprinkle', 'spray', 'splash',
'sprinkle1', 'spray1', 'and splash1',
}
\examples{
dontrun{data(toydata)
# run the appropriate function
result <- cheesepop(df = toydata$admin, output_dir) # run cheesepop or cheesecake
rclass <- paste0("TOY_population_v1_0_age",1:12) # specify the names of the raster files to be saved
class <- names(result$full_data \%>\% dplyr::select(starts_with("age_"))) # specify the raster files to visualise
output_dir <- paste0(out_path, "/raster/spray2") # set output directory
result2b <- spray(df=result$full_data, rdf=toydata$grid,
                rclass, output_dir)
#make raster maps
  #- use below to see the list of raster files saved in the directory
list.files(output_dir, pattern = "\\\\.tif$", full.names = TRUE)

group <- 1:12 # customised group #
raster_files <- paste0(tempdir(), "/pop_",paste0("TOY_population_v1_0_age",group), ".tif")
names <- paste0("Age ", group) # Customised names of the plot panels (same length as rclass)
nrow = 4; ncol = 3 # rows and columns of the panels of the output maps
plotRast(tempdir(), names, nrow, ncol)} # make the maps
}
